% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swmm-package.R
\docType{package}
\name{swmm-package}
\alias{swmm}
\alias{swmm-package}
\title{swmm: Low-level Access to the USEPA Stormwater Management Model (SWMM)}
\description{
Provides low-level Access to the USEPA Stormwater Management Model (SWMM) binaries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paleolimbot/swmm}
  \item Report bugs at \url{https://github.com/paleolimbot/swmm/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net}

}
\keyword{internal}
