% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swmm_error_messages}
\alias{swmm_error_messages}
\title{SWMM error messages}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 92 rows and 2 columns.}
\usage{
swmm_error_messages
}
\description{
Error messages table
}
\keyword{datasets}
