% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_files.R
\name{swmm_example_file}
\alias{swmm_example_file}
\alias{swmm_example_files}
\alias{swmm_example_dir}
\title{Get the path to an example input file}
\usage{
swmm_example_file(file)

swmm_example_files()

swmm_example_dir()
}
\arguments{
\item{file}{A file such as one listed by `swmm_example_files()`.}
}
\description{
Get the path to an example input file
}
\examples{
swmm_example_dir()
swmm_example_files()
swmm_example_file("Example1-Pre.inp")

}
