% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swmm_read_out.R
\name{swmm_out_variables}
\alias{swmm_out_variables}
\title{Get out file variable information}
\usage{
swmm_out_variables(pollutant_names = character(0))
}
\arguments{
\item{pollutant_names}{Optional names of pollutants in the system}
}
\value{
A tibble
}
\description{
Get out file variable information
}
\examples{
swmm_out_variables()

}
