% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swmm_read_out.R
\name{swmm_read_out}
\alias{swmm_read_out}
\title{Read SWMM out files in long format}
\usage{
swmm_read_out(path, element_type = NULL, variable = NULL, object_name = NULL)
}
\arguments{
\item{path}{A path to a .out file generated by \code{\link[=swmm_run]{swmm_run()}}.}

\item{element_type}{One of 'subcatchments', 'links', 'nodes', or 'system',
or \code{NULL} to read everything.}

\item{variable}{See \code{\link[=swmm_read_variables]{swmm_read_variables()}} for choices,
or \code{NULL} to read everything.}

\item{object_name}{An object name or \code{NULL} to read everything.}
}
\value{
A data frame
}
\description{
Read SWMM out files in long format
}
