% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swmm_read_out.R
\name{swmm_read_variables}
\alias{swmm_read_variables}
\alias{swmm_read_subcatchments}
\alias{swmm_read_links}
\alias{swmm_read_nodes}
\alias{swmm_read_system}
\title{Read .out file convenience functions}
\usage{
swmm_read_variables(path)

swmm_read_subcatchments(path, variable = NULL, object_name = NULL)

swmm_read_links(path, variable = NULL, object_name = NULL)

swmm_read_nodes(path, variable = NULL, object_name = NULL)

swmm_read_system(path, variable = NULL, object_name = NULL)
}
\arguments{
\item{path}{A path to a .out file generated by \code{\link[=swmm_run]{swmm_run()}}.}

\item{variable}{See \code{\link[=swmm_read_variables]{swmm_read_variables()}} for choices,
or \code{NULL} to read everything.}

\item{object_name}{An object name or \code{NULL} to read everything.}
}
\description{
Read .out file convenience functions
}
