% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swmmbin-package.R
\docType{package}
\name{swmmbin-package}
\alias{swmmbin}
\alias{swmmbin-package}
\title{swmmbin: Low-level Access to EPA Stormwater Management Model (SWMM)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net}

}
\keyword{internal}
