% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swmm-package.R
\docType{package}
\name{swmm-package}
\alias{swmm}
\alias{swmm-package}
\title{swmm: Cross-platform Access to the USEPA Stormwater Management Model (SWMM)}
\description{
Provides cross-platform access to the USEPA Stormwater Management Model (SWMM).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paleolimbot/swmm}
  \item Report bugs at \url{https://github.com/paleolimbot/swmm/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (0000-0002-9415-4582)

}
\keyword{internal}
