% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swmm_run.R
\name{swmm_run}
\alias{swmm_run}
\title{Initiate a simulation run}
\usage{
swmm_run(inp, rpt = NULL, out = NULL, overwrite = FALSE,
  quiet = TRUE)
}
\arguments{
\item{inp}{Path to an input file.}

\item{rpt}{Path to a report file (that does not yet exist or will be overwritten).}

\item{out}{Path to an output file (that does not yet exist or will be overwritten).}

\item{overwrite}{Use `overwrite = TRUE` overwrite `rpt` and `out`.}

\item{quiet}{Silence default SWWM output.}
}
\description{
This is intended to be a drop-in replacement for [swmmr::run_swmm()].
}
\examples{
swmm_run(swmm_example_file("Example1-Pre.inp"))

}
