% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swmmbin-package.R
\docType{package}
\name{swmmbin-package}
\alias{swmmbin}
\alias{swmmbin-package}
\title{swmmbin: Low-level Access to the USEPA Stormwater Management Model (SWMM)}
\description{
Provides low-level Access to the USEPA Stormwater Management Model (SWMM) binaries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paleolimbot/swmmbin}
  \item Report bugs at \url{https://github.com/paleolimbot/swmmbin/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net}

}
\keyword{internal}
